+++ Content/LeagueSandbox-Scripts/Buffs/Ascension/AscRespawn.cs
@@ -3,8 +3,11 @@ using GameServerCore.Enums;
 using GameServerCore.Domain.GameObjects.Spell;
 using static LeagueSandbox.GameServer.API.ApiFunctionManager;
 using static LeagueSandbox.GameServer.API.ApiMapFunctionManager;
+using LeagueSandbox.GameServer.API;
 using GameServerCore.Scripting.CSharp;
 using LeagueSandbox.GameServer.Scripting.CSharp;
+using GameServerCore.Domain;
+using LeagueSandbox.GameServer.GameObjects.Stats;
 
 namespace Buffs
 {
@@ -15,56 +18,28 @@ namespace Buffs
             BuffType = BuffType.INTERNAL,
             BuffAddType = BuffAddType.REPLACE_EXISTING
         };
-        public IStatsModifier StatsModifier { get; private set; }
+        public IStatsModifier StatsModifier { get; private set; } = new StatsModifier();
 
-        float soundTimer = 1000.0f;
-        IAttackableUnit Unit;
-        bool hasNotifiedSound = false;
         public void OnActivate(IAttackableUnit unit, IBuff buff, ISpell ownerSpell)
         {
-            Unit = unit;
-
-            if (unit is IObjAiBase obj)
+            if(unit is IObjAiBase obj && obj.Inventory != null)
             {
-                NotifyWorldEvent(EventID.OnMinionAscended);
-                NotifyAscendant(obj);
+                AddBuff("AscTrinketStartingCD", 0.2f, 1, null, unit, obj);
+                ApiEventManager.OnResurrect.AddListener(this, obj, OnRespawn, false);
             }
+        }
 
-            //Using SetStatus temporarily due to: https://github.com/LeagueSandbox/GameServer/issues/1385
-            SetStatus(unit, StatusFlags.Targetable, false);
-            SetStatus(unit, StatusFlags.Stunned, true);
-            SetStatus(unit, StatusFlags.Invulnerable, true);
-
-            unit.PauseAnimation(true);
-
-            AddParticleTarget(unit, unit, "EggTimer", unit, buff.Duration, flags: (FXFlags)32);
-            AddParticle(unit, unit, "AscTransferGlow", unit.Position, buff.Duration, flags: (FXFlags)32);
-            AddParticle(unit, unit, "AscTurnToStone", unit.Position, buff.Duration, flags: (FXFlags)32);
+        public void OnRespawn(IObjAiBase owner)
+        {
+            owner.Spells[6 + (byte)SpellSlotType.InventorySlots].SetCooldown(0, true);
         }
 
         public void OnDeactivate(IAttackableUnit unit, IBuff buff, ISpell ownerSpell)
         {
-            SetStatus(unit, StatusFlags.Targetable, true);
-            SetStatus(unit, StatusFlags.Stunned, false);
-            SetStatus(unit, StatusFlags.Invulnerable, false);
-
-            unit.PauseAnimation(false);
-            if (unit is IObjAiBase obj)
-            {
-                AddBuff("AscBuffIcon", 25000.0f, 1, null, obj, obj);
-                AddBuff("AscXerathControl", 999999.0f, 1, null, obj, obj);
-            }
         }
 
         public void OnUpdate(float diff)
         {
-            soundTimer -= diff;
-            if (soundTimer <= 0 && !hasNotifiedSound)
-            {
-                PlaySound("Play_sfx_ZhonyasRingShield_OnBuffActivate", Unit);
-                PlaySound("Play_sfx_Cassiopeia_CassiopeiaPetrifyingGazeStun_OnBuffActivate", Unit);
-                hasNotifiedSound = true;
-            }
         }
     }
 }
