+++ Content/LeagueSandbox-Scripts/Maps/Map8/LevelScriptAscension.cs
@@ -117,6 +117,7 @@ namespace MapScripts.Map8
             {
                 ApiEventManager.OnIncrementChampionScore.AddListener(this, champion, OnIncrementPoints, false);
                 ApiEventManager.OnKill.AddListener(this, champion, OnChampionKill, false);
+                AddBuff("AscRespawn", 25000.0f, 1, null, champion, champion);
             }
 
             foreach (var player in GetAllPlayers())
